Added user information to login request
