+++ backend/accounts/jwt_handler.py
@@ -1,0 +1,7 @@
+from accounts.serializer import UserSerializer
+
+def jwt_response_payload_handler(token, user=None, request=None):
+    return {
+        'token': token,
+        'user': UserSerializer(user, context={'request': request}).data
+    }
