+++ backend/accounts/views.py
@@ -1,21 +1,11 @@
 from django.contrib.auth.models import User
-from rest_framework import serializers
+from accounts.serializer import UserSerializer, UserSerializerWithPassword
 from rest_framework import generics
 from rest_framework.permissions import AllowAny
-
-class UserSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = User
-        fields = ('id', 'username','password', 'email', 'first_name', 'last_name')
-
-    def create(self, validated_data):
-        user = User.objects.create_user(**validated_data)
-        return user
 
 
 
 class UserRegister(generics.CreateAPIView):
     permission_classes = (AllowAny,)
     queryset = User.objects.all()
-    serializer_class = UserSerializer
+    serializer_class = UserSerializerWithPassword
